requirements.txt
opencv-contrib-python==4.5.5.62
pillow==9.3.0
pytesseract==0.3.10
numpy~=1.23.5
scipy~=1.9.3
imutils~=0.5.4